requirements.txt
Django>=4.0.0,<5.0.0
djangorestframework>=3.12.4,<4.0.0
dj-rest-auth>=2.1.11,<3.0.0
djangorestframework-simplejwt>=5.0.0,<6.0.0
django-environ>=0.8.1,<1.0.0
django-extensions>=3.1.5,<4.0.0
psycopg2>=2.9.2,<3.0.0
bcrypt>=3.2.0,<4.0.0

flake8>=4.0.1,<5.0.0